+++ src/pub/proxy_edit/rules.js
@@ -26,7 +26,7 @@ const trigger_types = [
         pulled when the response <body> contain the selected string`},
     {key: 'Request time more than', value: 'min_req_time',
         tooltip: `Triggers when the request time is above the selected value`,
-        type: 'pre'},
+        type: '--pre'},
     {key: 'Request time less than', value: 'max_req_time',
         tooltip: `Triggers when the request time is below the selected value`},
 ];
@@ -36,12 +36,13 @@ const default_action = {key: 'Choose an action type', value: '',
     action is executed automatically.`};
 const action_types = [
     {key: 'Retry with new IP', value: 'retry', tooltip: `System will send the
-        exact same request again with newly refreshed IP`, min_req_time: true},
+        exact same request again with newly refreshed IP`,
+        min_req_time: false},
     {key: 'Retry with new proxy port (Waterfall)', value: 'retry_port',
         tooltip: `System will send another request using different port
         from your port list. This can allow cost optimization by escalating the
         request between different types of networks according to the port
-        configuration.`, min_req_time: true},
+        configuration.`, min_req_time: false},
     {key: 'Ban IP', value: 'ban_ip', tooltip: `Will ban the IP for custom
         amount of time. usually used for failed request.`},
     {key: 'Refresh IP', value: 'refresh_ip', tooltip: `Refresh the current
@@ -393,9 +394,9 @@ const Rule = with_proxy_ports(withRouter(class Rule extends Pure_component {
         .filter(at=>at.value!='save_to_fast_pool'||
             rule.trigger_type=='max_req_time')
         .filter(at=>rule.trigger_type=='url'&&at.only_url||
-            rule.trigger_type!='url'&&!at.only_url))
-        .filter(at=>rule.trigger_type!='min_req_time'||
-            at.min_req_time);
+            rule.trigger_type!='url'&&!at.only_url));
+        // .filter(at=>rule.trigger_type!='min_req_time'||
+        //     at.min_req_time);
         const ports = this.props.ports_opt.filter(p=>
             p.value!=this.props.match.params.port);
         return <div className="rule_wrapper">