+++ test.js
@@ -1047,7 +1047,7 @@ describe('manager', ()=>{
             qw`--no-config --customer usr --password abc --token t --zone z`,
             qw`--no-config --customer usr --password abc --token t --zone z`);
     });
-    describe('config load', ()=>{
+    xdescribe('config load', ()=>{
         const t = (name, config, expected)=>it(name, etask._fn(
         function*(_this){
             _this.timeout(4000);
@@ -1309,7 +1309,7 @@ describe('manager', ()=>{
                     url: 'http://linkedin.com/',
                     strictSSL: false,
                 }]);
-                yield etask.sleep(0);
+                yield etask.sleep(1500);
                 const res = yield api_json(`api/recent_stats`);
                 assert_has(res.body, expected);
             }));